Conflicts:
	source/Genome.h
	source/Parameters.cpp
	source/Parameters.h
	source/ReadAlign.cpp
	source/ReadAlign_alignBAM.cpp
	source/ReadAlign_quantTranscriptome.cpp
	source/STAR.cpp
	source/Transcript.h
	source/loadGTF.cpp
	source/parametersDefault.xxd
	source/serviceFuns.cpp
	source/sjdbPrepare.cpp
	source/stitchWindowAligns.cpp

+++ source/Transcriptome_geneCountsAddAlign.cpp
@@ -11,23 +11,23 @@ void Transcriptome::geneCountsAddAlign(uint nA, Transcript **aAll) {//add alignm
 
          int64 e1=-1;
 
-         for (int ib=a.nExons-1; ib>=0; ib--) 
+         for (int ib=a.nExons-1; ib>=0; ib--)
          {//scan through all blocks of the alignments
 
              uint64 g1=a.exons[ib][EX_G]+a.exons[ib][EX_L]-1;//end of the block
 
-//              if ((uint)ib==a.nExons-1) 
+//              if ((uint)ib==a.nExons-1)
 //              {//binary search for the first time: end of the block among the starts of exons
                  e1=binarySearch1a<uint64>(g1, exG.s, (int32) exG.nEx);
-//              } else 
-//              {//simple backwards scan 
-//                  while (e1>=0 && exG.s[e1]>g1) 
+//              } else
+//              {//simple backwards scan
+//                  while (e1>=0 && exG.s[e1]>g1)
 //                  {//stop when exon start is less than block end
 //                      --e1;
 //                  };
 //              };
 
-             while (e1>=0 && exG.eMax[e1]>=a.exons[ib][EX_G]) 
+             while (e1>=0 && exG.eMax[e1]>=a.exons[ib][EX_G])
              {//these exons may overlap this block
                  if (exG.e[e1]>=a.exons[ib][EX_G])
                  {//this exon overlaps the block
@@ -38,35 +38,35 @@ void Transcriptome::geneCountsAddAlign(uint nA, Transcript **aAll) {//add alignm
                          if ( itype==1 && a.Str!=str1 && str1<2) continue; //same strand
                          if ( itype==2 && a.Str==str1 && str1<2) continue; //reverse strand
 
-                         if (gene1.at(itype)==-1) 
+                         if (gene1.at(itype)==-1)
                          {//first gene overlapping this read
                              gene1[itype]=exG.g[e1];
-                         } else if (gene1.at(itype)==-2) 
+                         } else if (gene1.at(itype)==-2)
                          {
                              continue;//this align was already founf to be ambig for this strand
-                         } else if (gene1.at(itype)!=exG.g[e1]) {//another gene overlaps this read
+                         } else if (gene1.at(itype)!=(int32)exG.g[e1]) {//another gene overlaps this read
                              gene1[itype]=-2;//mark ambiguous
                          };//otherwise it's the same gene
                      };
                  };
-                 
+
                  --e1;// go to the previous exon
              };
          };
 
-        for (int itype=0; itype<quants->geneCounts.nType; itype++)         
+        for (int itype=0; itype<quants->geneCounts.nType; itype++)
         {
-            if (gene1.at(itype)==-1) 
+            if (gene1.at(itype)==-1)
             {
                 quants->geneCounts.cNone[itype]++;
-            } else if (gene1.at(itype)==-2) 
+            } else if (gene1.at(itype)==-2)
             {
                 quants->geneCounts.cAmbig[itype]++;
             } else
             {
                 quants->geneCounts.gCount[itype][gene1.at(itype)]++;
             };
-         }; 
+         };
      };
 };
